// src/app/modules/academicSemester/academicSemester.service.ts
import { AcademicSemesterNameCodeMapper } from './academicSemester.constant';
import { TAcademicSemester } from './academicSemester.interface';
import { AcademicSemester } from './academicSemester.model';

const createAcademicSemesterIntoDb = async (payload: TAcademicSemester) => {
  // Semester name ---> semester code

  // type TAcademicSemesterNameCodeMapper = {
  //   Autumn: '01';
  //   Summar: '02';
  //   Fall: '03';
  // };

  //AcademicSemesterNameCodeMapper['Fall'] ->03
  if (AcademicSemesterNameCodeMapper[payload.name] !== payload.code) {
    throw new Error('Invalid semester code!!');
  }

  const result = await AcademicSemester.create(payload);
  return result;
};

const getAllAcademicSemesterFromDB = async () => {
  const result = await AcademicSemester.find();
  return result;
};

const getSingleAcademicSemesterFromDB = async (id: string) => {
  //const result = await AcademicSemester.findOne({ id });
  const result = await AcademicSemester.findById( id );
  //const result = await AcademicSemester.aggregate([{ $match: { id: id } }]);
  return result;
};

const updateAcademicSemesterIntoDB = async (
  id: string,
  payload: Partial<TAcademicSemester>,
) => {
  if (
    payload.name &&
    payload.code &&
    AcademicSemesterNameCodeMapper[payload.name] !== payload.code
  ) {
    throw new Error('Invalid Semester Code');
  }

  const result = await AcademicSemester.findOneAndUpdate({ _id: id }, payload, {
    new: true,
  });
  return result;
};


export const AcademicSemesterService = {
  createAcademicSemesterIntoDb,
  getAllAcademicSemesterFromDB,
  getSingleAcademicSemesterFromDB,
  updateAcademicSemesterIntoDB
};
